Add files via upload
A project that uses reverse polish notation for a calculator, parses command line arguments, converts those arguments from strings-to-doubles, and performs operations upon numbers after being pushed onto a stack. Please note that division is NOT included in this code, and that built-in functions aren't allowed, (strtod / sscanf and etc.).

Difficulty from 1 - 5 : 4.